refactor(product): Improve validation and remove sizes/colors from entity
- Updated validation logic in ProductRequest
- Removed `sizes` and `colors` fields from Product entity
- Refactored related methods to align with new product structure
- Ensured smooth migration and compatibility with existing API flows
