+++ src/main/java/com/gotrid/trid/api/product/dto/ProductResponse.java
@@ -1,12 +1,10 @@
-package com.gotrid.trid.api.shop.dto.product;
+package com.gotrid.trid.api.product.dto;
 
 import java.math.BigDecimal;
 
 public record ProductResponse(
         Integer id,
         String name,
-        String sizes,
-        String colors,
         String description,
         BigDecimal basePrice
 ) {
