+++ src/main/java/com/gotrid/trid/api/shop/dto/ShopRequest.java
@@ -1,4 +1,4 @@
-package com.gotrid.trid.api.shop.dto.shop;
+package com.gotrid.trid.api.shop.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
