+++ src/main/java/com/gotrid/trid/api/shop/service/ShopService.java
@@ -1,19 +1,23 @@
 package com.gotrid.trid.api.shop.service;
 
-import com.gotrid.trid.api.shop.dto.CoordinateDTO;
+import com.gotrid.trid.api.product.service.ProductService;
+import com.gotrid.trid.core.threedModel.dto.CoordinateDTO;
 import com.gotrid.trid.api.shop.dto.SocialDTO;
-import com.gotrid.trid.api.shop.dto.shop.ShopModelResponse;
-import com.gotrid.trid.api.shop.dto.shop.ShopRequest;
-import com.gotrid.trid.api.shop.dto.shop.ShopResponse;
-import com.gotrid.trid.api.shop.dto.shop.ShopUpdateRequest;
+import com.gotrid.trid.api.shop.dto.ShopModelResponse;
+import com.gotrid.trid.api.shop.dto.ShopRequest;
+import com.gotrid.trid.api.shop.dto.ShopResponse;
+import com.gotrid.trid.api.shop.dto.ShopUpdateRequest;
 import com.gotrid.trid.common.exception.custom.shop.ShopException;
 import com.gotrid.trid.common.exception.custom.shop.ShopNotFoundException;
 import com.gotrid.trid.common.response.PageResponse;
-import com.gotrid.trid.core.shop.mapper.CoordinateMapper;
+import com.gotrid.trid.core.product.model.Product;
+import com.gotrid.trid.core.threedModel.mapper.CoordinateMapper;
 import com.gotrid.trid.core.shop.mapper.ShopMapper;
 import com.gotrid.trid.core.shop.mapper.SocialMapper;
 import com.gotrid.trid.core.shop.model.*;
 import com.gotrid.trid.core.shop.repository.ShopRepository;
+import com.gotrid.trid.core.threedModel.model.Coordinates;
+import com.gotrid.trid.core.threedModel.model.Model;
 import com.gotrid.trid.core.user.model.Users;
 import com.gotrid.trid.core.user.repository.UserRepository;
 import com.gotrid.trid.infrastructure.azure.ShopStorageService;
