+++ src/main/java/com/gotrid/trid/core/product/model/Product.java
@@ -1,6 +1,8 @@
-package com.gotrid.trid.core.shop.model;
+package com.gotrid.trid.core.product.model;
 
 import com.gotrid.trid.common.model.AuditableEntity;
+import com.gotrid.trid.core.threedModel.model.Model;
+import com.gotrid.trid.core.shop.model.Shop;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
@@ -20,10 +22,6 @@ public class Product extends AuditableEntity {
 
     @Column(length = 100)
     private String name;
-    @Column(length = 50)
-    private String sizes;
-    @Column(length = 50)
-    private String colors;
     @Column(length = 1000)
     private String description;
     @Column(precision = 10, scale = 2)
