+++ src/main/java/com/gotrid/trid/core/product/repository/ProductVariantRepository.java
@@ -1,6 +1,6 @@
-package com.gotrid.trid.core.shop.repository;
+package com.gotrid.trid.core.product.repository;
 
-import com.gotrid.trid.core.shop.model.ProductVariant;
+import com.gotrid.trid.core.product.model.ProductVariant;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
