+++ src/main/java/com/gotrid/trid/core/threedModel/dto/ModelDTO.java
@@ -1,4 +1,4 @@
-package com.gotrid.trid.api.shop.dto;
+package com.gotrid.trid.core.threedModel.dto;
 
 public record ModelDTO(
         String glbUrl,
