+++ src/main/java/com/gotrid/trid/core/threedModel/repository/CoordinatesRepository.java
@@ -1,6 +1,6 @@
-package com.gotrid.trid.core.shop.repository;
+package com.gotrid.trid.core.threedModel.repository;
 
-import com.gotrid.trid.core.shop.model.Coordinates;
+import com.gotrid.trid.core.threedModel.model.Coordinates;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CoordinatesRepository extends JpaRepository<Coordinates, Integer> {
