+++ src/main/resources/db/migration/V13__remove_sizes_and_colors_from_product.sql
@@ -1,0 +1,3 @@
+ALTER TABLE product
+    DROP COLUMN sizes,
+    DROP COLUMN colors;
